refactor: migrate from Cassandra to PostgreSQL
Replace Apache Cassandra with PostgreSQL for all data storage. The workload
is relational (dual writes for bidirectional lookups, ALLOW FILTERING
anti-patterns, no JOINs) and doesn't need Cassandra's distributed scale.

Key changes:
- sqlx with PgPool replaces scylla driver (Cargo.toml, db.rs, auth_db.rs)
- Row-Level Security (RLS) replaces per-keyspace tenant isolation
- FOR UPDATE SKIP LOCKED replaces poll+delete queue pattern
- Surrogate UUIDs (page_id, queue_id, error_id) on core tables
- Soft deletion (deleted_at) on tenants, users, domains, pages, memberships
- Single org_memberships table replaces dual org_memberships + user_orgs
- Aggregation queries replace crawl_stats counter table
- docker/postgres/schema.sql replaces docker/cassandra/ directory
- All docs, tests, CI, and scripts updated

+++ lala-agent/src/models/storage.rs
@@ -27,17 +27,16 @@ impl CompressionType {
         }
     }
 
-    /// Convert to database tinyint representation
-    pub fn to_db_value(&self) -> i8 {
+    /// Convert to database SMALLINT representation
+    pub fn to_db_value(&self) -> i16 {
         match self {
             CompressionType::None => 0,
             CompressionType::Gzip => 1,
         }
     }
 
-    /// Parse from database tinyint representation
-    /// Handles NULL values (from rows created before the column existed)
-    pub fn from_db_value(value: Option<i8>) -> Self {
+    /// Parse from database SMALLINT representation
+    pub fn from_db_value(value: Option<i16>) -> Self {
         match value {
             Some(1) => CompressionType::Gzip,
             _ => CompressionType::None,
